backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
kafka-python==2.0.2
